pyproject.toml
[project]
name = "db-work-template"
version = "0.1.0"
description = ""
authors = [
    {name = "rodrigoneal",email = "rodrigho2006@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "sqlmodel (>=0.0.24,<0.0.25)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pymysql (>=1.1.1,<2.0.0)",
    "sqlalchemy[asyncio] (>=2.0.41,<3.0.0)",
    "aiomysql (>=0.2.0,<0.3.0)"
]

[tool.poetry]
packages = [{include = "db_work_template", from = "src"}]


[tool.poetry.group.dev.dependencies]
isort = "^6.0.1"
ruff = "^0.12.4"
groups = "^0.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
